rename: planner-pannel 폴더 이름 변경
