+++ src/app/workspace/[id]/page.tsx
@@ -1,6 +1,6 @@
 // 작업공간
 import MemoPannel from './components/memo-pannel/MemoPannel'
-import PlannerPannel from './components/planner-panel/PlannerPannel'
+import PlannerPannel from './components/planner-pannel/PlannerPannel'
 
 export default function Page() {
   return (
